PhySO-Telegram-Bot/requirements.txt
python-telegram-bot==20.7
requests>=2.28.0
pandas>=1.5.0
aiofiles>=23.0.0
python-dotenv>=1.0.0
